Improve TT entry move updating
Do move-only update if we have a new position or if the position
already didn't have a move. Also, don't reset the move for existing
positions. This fixes a case where the search may take an excessive
amount of time until finding the correct move.

To reproduce:

Use parent commit: 530fccbf272ffe424ae53a464b91db148cc968ae

  setoption name Contempt value 0
  setoption name Hash value 32768
  setoption name MultiPV value 1
  position fen r1bqk1nr/pp2bppp/2np4/2p1p1N1/2B1P3/2NP4/PPP2PPP/R1BQK2R b KQkq - 2 6
  setoption name UCI_AnalyseMode value true
  go infinite

Without this patch, it takes 57M nodes for the search to find Bg5 with
cp >100 score. With this patch, it takes only 18M nodes.

Bench: 3890393